// src/protobuf/feoblog.ts
export type Item = Message<"Item"> & {
  /**
   * REQUIRED
   * The timestamp is used to give order to a user's collection of Items.
   * This timestamp represents the number of milliseconds since
   * 1970-01-01 00:00:00.000Z (ignoring leap seconds).
   * 
   * This is somewhat useful for displaying blog posts in order. But it's
   * especially important for ordering things like updates to a user's
   * profile.
   * 
   * As a result, servers should not accept timestamps in the future (except
   * for maybe a small allowance due to clock drift.)
   * 
   * Servers must validate that this value is specified.
   * 
   * Due to protobuf3 default values, this means the value
   * can not be exactly 0.
   * Update: With new versionf of proto3, we can now unambiguously specify 0.
   * TODO: Implement that.
   *
   * @generated from field: int64 timestamp_ms_utc = 1;
   */
  timestampMsUtc: bigint;

  /**
   * Optionally specify the user's timezone offset when they created this
   * Item. This is useful when displaying more meaningful dates on things
   * like Posts.
   * 
   * For example, Pacific Daylight Time has an offset of "-0700", or -420
   * minutes.
   * Servers should reject offsets of more than +/- 24 hours. 
   * 
   * Defaults to 0 (UTC).
   *
   * @generated from field: sint32 utc_offset_minutes = 2;
   */
  utcOffsetMinutes: number;

  /**
   * @generated from oneof Item.item_type
   */
  itemType: {
    /**
     * @generated from field: Post post = 3;
     */
    value: Post;
    case: "post";
  } | {
    /**
     * @generated from field: Profile profile = 4;
     */
    value: Profile;
    case: "profile";
  } | {
    /**
     * @generated from field: Comment comment = 5;
     */
    value: Comment;
    case: "comment";
  } | { case: undefined; value?: undefined };
};
// ...
export type Post = Message<"Post"> & {
  /**
   * An optional plaintext title for the post.
   * Titles should be <= 256 bytes. Servers may reject longer ones.
   *
   * @generated from field: string title = 1;
   */
  title: string;

  /**
   * The body of the post, formatted in CommonMark markdown.
   * Servers should suppress unsafe raw HTML blocks in the body. They may do
   * so by rejecting the Item at the time of upload, or by choosing to render
   * the Item without the offending HTML parts.
   * 
   * The allowed size of the body is effectively limited by the allowed
   * size of the enclosing Item.
   *
   * @generated from field: string body = 2;
   */
  body: string;

  /**
   * File attachments that will be visible at ./files/*
   *
   * @generated from field: Attachments attachments = 5;
   */
  attachments?: Attachments;
};
// ...
export type Profile = Message<"Profile"> & {
  /**
   * A name to display instead of your userID.
   *
   * @generated from field: string display_name = 1;
   */
  displayName: string;

  /**
   * An "about me" section, formatted in Commonmark markdown.
   * Servers should suppress unsafe raw HTML blocks in the body.
   *
   * @generated from field: string about = 2;
   */
  about: string;

  /**
   * A list of servers where the user expects their content to be hosted.
   * The first server is considered the "primary" server, but others may be listed
   * as backups.
   * This allows users to move servers by updating their preferred server list.
   *
   * @generated from field: repeated Server servers = 3;
   */
  servers: Server[];

  /**
   * A list of users who this user "follows".
   * This allows the server to know what additional users it should cache data for, so that it can present this
   * (Profile) user's feed of new content.
   * 
   * The order of the list is unimportant.
   *
   * @generated from field: repeated Follow follows = 4;
   */
  follows: Follow[];

  /**
   * Users may group their follows into groups to make sorting/filtering/syncing them easier.
   * Clients may decide to order groups in different ways, or respect the order of the groups as listed here.
   *
   * @generated from field: repeated FollowGroup follow_groups = 5;
   */
  followGroups: FollowGroup[];
};

/**
 * Describes the message Profile.
 * Use `create(ProfileSchema)` to create a new message.
 */
export const ProfileSchema: GenMessage<Profile> = /*@__PURE__*/
  messageDesc(file_feoblog, 2);

/**
 * A Comment is a text-only response to some other Item.
 *
 * @generated from message Comment
 */
export type Comment = Message<"Comment"> & {
  /**
   * Information about the Item we're replying to.
   *
   * @generated from field: ReplyRef reply_to = 1;
   */
  replyTo?: ReplyRef;

  /**
   * CommonMark markdown text. 
   * Inline images will NOT be rendered.
   *
   * @generated from field: string text = 2;
   */
  text: string;
};

/**
 * Describes the message Comment.
 * Use `create(CommentSchema)` to create a new message.
 */
export const CommentSchema: GenMessage<Comment> = /*@__PURE__*/
  messageDesc(file_feoblog, 3);

/**
 * Information about an Item that we're replying to.
 *
 * @generated from message ReplyRef
 */
export type ReplyRef = Message<"ReplyRef"> & {
  /**
   * REQUIRED: the user_id that posted the item.
   *
   * @generated from field: UserID user_id = 1;
   */
  userId?: UserID;

  /**
   * REQUIRED: the signature of the item.
   *
   * @generated from field: Signature signature = 2;
   */
  signature?: Signature;

  /**
   * Suggested: The type of the item we're replying to.
   * This is useful in case clients want to only show certain types of comments. 
   * Ex: show comments in reply to posts, but not comments replying to other comments.
   *
   * @generated from field: ItemType item_type = 3;
   */
  itemType: ItemType;
};
// ...
export type Server = Message<"Server"> & {
  /**
   * A URL to a FeoBlog server.
   * Ex:
   * "https://feo.example.com"
   * "https://feo.example.com/"
   * "https://feo.example.com:8080"
   * "https://feo.example.com:8080/"
   * 
   * Note 1: Subpaths are currently not supported. Ex: "https://feo.example.com/some/subpath/"
   * 
   * Note 2: While the signature authentication allows serving/sending signed protobuf Items
   * securely without HTTPS, since the main client is currently implemented as in-browser JavaScript,
   * it is recommended that servers use only HTTPS to avoid JavaScript injection in the client.
   * So, you should probably prefer https for the REST endpoints too.
   * Plus, HTTP/2 requires HTTPS, and you'll get better performance for fetching many small Items w/ HTTP/2.
   *
   * @generated from field: string url = 1;
   */
  url: string;
};

/**
 * Describes the message Server.
 * Use `create(ServerSchema)` to create a new message.
 */
export const ServerSchema: GenMessage<Server> = /*@__PURE__*/
  messageDesc(file_feoblog, 5);

/**
 * @generated from message Follow
 */
export type Follow = Message<"Follow"> & {
  /**
   * REQUIRED
   *
   * @generated from field: UserID user = 1;
   */
  user?: UserID;

  /**
   * Set a display name for a user within the context of your feed.
   * 
   * Users may change their display names in their profiles. But, unlike Twitter, FeoBlog does not have
   * a globally-unique human-readable ID to fall back on to identify someone, so it can be difficult to
   * know who's who if people keep changing their names. 
   * Here you can set a stable name so you always know who's who.
   *
   * @generated from field: string display_name = 2;
   */
  displayName: string;

  /**
   * The (0-indexed) group index that this follow should be grouped under.
   *
   * @generated from field: optional int32 follow_group = 3;
   */
  followGroup?: number;
};

/**
 * Describes the message Follow.
 * Use `create(FollowSchema)` to create a new message.
 */
export const FollowSchema: GenMessage<Follow> = /*@__PURE__*/
  messageDesc(file_feoblog, 6);

/**
 * @generated from message FollowGroup
 */
export type FollowGroup = Message<"FollowGroup"> & {
  /**
   * A display name for this group. ex: "News", or "Friends".
   * Names should probably be unique for users' own sanity but at the moment that is not enforced by the protocol.
   *
   * @generated from field: string name = 1;
   */
  name: string;
};

/**
 * Describes the message FollowGroup.
 * Use `create(FollowGroupSchema)` to create a new message.
 */
export const FollowGroupSchema: GenMessage<FollowGroup> = /*@__PURE__*/
  messageDesc(file_feoblog, 7);

/**
 * @generated from message UserID
 */
export type UserID = Message<"UserID"> & {
  /**
   * A user's public NaCL key/ID. Must be 32 bytes:
   *
   * @generated from field: bytes bytes = 1;
   */
  bytes: Uint8Array;
};

/**
 * Describes the message UserID.
 * Use `create(UserIDSchema)` to create a new message.
 */
export const UserIDSchema: GenMessage<UserID> = /*@__PURE__*/
  messageDesc(file_feoblog, 8);

/**
 * @generated from message Signature
 */
export type Signature = Message<"Signature"> & {
  /**
   * A NaCL signature. Must be 64 bytes:
   *
   * @generated from field: bytes bytes = 1;
   */
  bytes: Uint8Array;
};
// ...
export type ItemListEntry = Message<"ItemListEntry"> & {
  /**
   * user_id may be unspecified if it can be inferred from context.
   * (ex: in an ItemList which lists posts for a specific userID)
   *
   * @generated from field: UserID user_id = 1;
   */
  userId?: UserID;

  /**
   * REQUIRED.
   *
   * @generated from field: Signature signature = 2;
   */
  signature?: Signature;

  /**
   * REQUIRED
   * The timestamp coantained within Item.timestamp_ms_utc.
   * This is used for ordering Items, and to fetch more ItemIDs in the event
   * that this list is truncated/incomplete.   
   *
   * @generated from field: int64 timestamp_ms_utc = 3;
   */
  timestampMsUtc: bigint;

  /**
   * Specify the type of this item.
   * This allows clients to skip fetching item types they're not interested in
   * for a particular view. (ex: profile updates and/or comments, etc.)
   *
   * @generated from field: ItemType item_type = 4;
   */
  itemType: ItemType;
};

/**
 * Describes the message ItemListEntry.
 * Use `create(ItemListEntrySchema)` to create a new message.
 */
export const ItemListEntrySchema: GenMessage<ItemListEntry> = /*@__PURE__*/
  messageDesc(file_feoblog, 11);

/**
 * File attachments.
 * Certain item types may allow file attachments.
 *
 * @generated from message Attachments
 */
export type Attachments = Message<"Attachments"> & {
  /**
   * @generated from field: repeated File file = 1;
   */
  file: File[];
};
// ...
export enum ItemType {
  /**
   * Default value. Either the server didn't specify the type, or
   * it was a type that the client can't deserialize.
   *
   * @generated from enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * @generated from enum value: POST = 1;
   */
  POST = 1,

  /**
   * @generated from enum value: PROFILE = 2;
   */
  PROFILE = 2,

  /**
   * @generated from enum value: COMMENT = 3;
   */
  COMMENT = 3,
}